worker/requirements.txt
flask
flask_sqlalchemy
json
urlparse
urllib3
